requirements.txt
absl-py==0.8.1
albumentations==0.3.0
altgraph==0.16.1
apex==0.1
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
astropy==3.2.3
atomicwrites==1.3.0
attrs==19.1.0
audioread==2.1.6
autopep8==1.4.3
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
blis==0.2.4
Bottleneck==1.2.1
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.4.1
command-not-found==0.3
confuse==1.0.0
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.2
Cython==0.29.1
dask==1.0.0
dataclasses==0.6
dcase-util==0.2.10
decorator==4.3.0
defer==1.0.6
defusedxml==0.5.0
distro-info===0.18ubuntu0.18.04.1
docopt==0.6.2
easydict==1.9
efficientnet==1.0.0b3
efficientnet-pytorch==0.5.1
entrypoints==0.3
fastai==1.0.55
fastprogress==0.1.21
future==0.17.1
gast==0.2.0
google-images-download==2.8.0
google-pasta==0.1.8
graphviz==0.8.4
grpcio==1.16.1
h5py==2.8.0
htmlmin==0.1.12
httplib2==0.9.2
idna==2.6
image-classifiers==1.0.0b1
imageio==2.4.1
imgaug==0.2.6
importlib-metadata==0.23
iotop==0.6
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.21
jedi==0.13.2
Jinja2==2.10
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-rectified-adam==0.9.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
librosa==0.6.3
llvmlite==0.28.0
lmdb==0.98
louis==3.5.0
lxml==4.2.6
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.2
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==7.2.0
munch==2.3.2
murmurhash==1.0.2
mxnet-cu90==1.5.1.post0
nbconvert==5.4.0
nbformat==4.4.0
netifaces==0.10.4
networkx==2.2
notebook==5.7.4
numba==0.43.1
numexpr==2.6.9
numpy==1.16.5
nvidia-ml-py3==7.352.0
oauth==1.0.1
olefile==0.45.1
opencv-python==3.4.1.15
opencv-python-headless==4.1.0.25
packaging==19.0
pandas==0.24.2
pandas-profiling==2.3.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.2.1
phik==0.9.8
pickleshare==0.7.5
Pillow==6.1.0
pipdeptree==0.13.2
pipreqs==0.4.9
plac==0.9.6
pluggy==0.13.0
preshed==2.0.1
pretrainedmodels==0.7.4
progressbar2==3.38.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
py==1.8.0
pycairo==1.16.2
pycocotools==2.0.0
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pydensecrf==1.0rc3
pydot==1.4.0
pydot-ng==2.0.0
Pygments==2.3.1
pygobject==3.26.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.3.0
pyRFC3339==1.0
pytest==5.2.2
pytest-pylint==0.14.1
python-apt==1.6.3+ubuntu1
python-dateutil==2.7.5
python-debian==0.1.32
python-magic==0.4.15
python-utils==2.3.0
pytorch-ignite==0.1.2
pytz==2018.3
PyWavelets==1.0.1
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.3
reportlab==3.4.0
requests==2.20.1
requests-unixsocket==0.1.5
resampy==0.2.1
scikit-image==0.14.1
scikit-learn==0.20.3
scipy==1.1.0
screen-resolution-extra==0.0.0
seaborn==0.9.0
SecretStorage==2.3.1
segmentation-models==1.0.0b1
segmentation-models-pytorch==0.1.0
selenium==3.141.0
Send2Trash==1.5.0
shadowsocks==2.8.2
Shapely==1.6.4.post2
simplejson==3.13.2
six==1.11.0
SoundFile==0.10.2
soupsieve==1.9.2
spacy==2.1.6
srsly==0.0.7
system-service==0.3
systemd-python==234
tensorboard==1.12.2
tensorboardX==1.8
tensorflow-gpu==1.12.3
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thinc==7.0.8
tldr==0.5
toml==0.10.0
toolz==0.9.0
torch==1.1.0
torchfile==0.1.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.3.0
tornado==5.1.1
tqdm==4.36.1
traitlets==4.3.2
ttach==0.0.2
typed-ast==1.4.0
typing==3.7.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
validators==0.14.0
visdom==0.1.8.8
wadllib==1.3.2
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==0.56.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.2
xkit==0.0.0
xlrd==1.2.0
yacs==0.1.6
yarg==0.1.9
youtube-dl==2018.3.14
zipp==0.6.0
zope.interface==4.3.2
